Change data type of 3 member variables of 'Process' class

+++ Process.c
@@ -126,9 +126,9 @@ typedef struct Process_ {
    float percent_cpu;
    float percent_mem;
 
-   long int priority;
-   long int nice;
-   long int nlwp;
+   int priority;
+   int nice;
+   int nlwp;
    time_t starttime_ctime;
 
    // In pages
